# Agent-Study/src/main/resources/swagger/openAI-txt.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agent Study API",
    "description": "OpenAPI documentation powered by springdoc-openapi",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:6363",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/chat/persist": {
      "post": {
        "tags": [
          "聊天管理"
        ],
        "summary": "手动持久化对话",
        "description": "立即将 Redis 中的对话记录持久化到数据库",
        "operationId": "persistConversation",
        "parameters": [
          {
            "name": "conversationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/chat/tool": {
      "get": {
        "tags": [
          "聊天管理"
        ],
        "summary": "使用工具与模型聊天",
        "description": "向聊天模型发送消息，并使用工具进行辅助",
        "operationId": "chatWithTool",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 3600
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
    "/chat/tool/stream": {
      "get": {
        "tags": [
          "聊天管理"
        ],
        "summary": "使用工具与模型聊天（SSE流式）",
        "description": "向聊天模型发送消息，使用工具辅助，通过SSE流式返回响应",
        "operationId": "chatWithToolStream",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 3600
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
    "/chat/push": {
      "get": {
        "tags": [
          "聊天管理"
        ],
        "summary": "使用记忆与模型聊天",
        "description": "向聊天模型发送消息，并根据聊天内存检索响应",
        "operationId": "chat",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 3600
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
    "/chat/push/stream": {
      "get": {
        "tags": [
          "聊天管理"
        ],
        "summary": "使用记忆与模型聊天（SSE流式）",
        "description": "向聊天模型发送消息，通过SSE流式返回响应",
        "operationId": "chatStream",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 3600
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
    "/chat/mcp": {
      "get": {
        "tags": [
          "聊天管理"
        ],
        "summary": "使用MCP工具与模型聊天",
        "description": "向聊天模型发送消息，并使用MCP工具进行辅助",
        "operationId": "chatWithMcpTool",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 3600
            }
          },
          {
            "name": "mcpNames",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
    "/chat/mcp/stream": {
      "get": {
        "tags": [
          "聊天管理"
        ],
        "summary": "使用MCP工具与模型聊天（SSE流式）",
        "description": "向聊天模型发送消息，使用MCP工具辅助，通过SSE流式返回响应",
        "operationId": "chatWithMcpToolStream",
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expireSeconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 3600
            }
          },
          {
            "name": "mcpNames",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultBean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "返回信息"
          },
          "data": {
            "type": "object",
            "description": "返回数据"
          }
        }
      },
...
    }
  }
}